perf(web): isolate logo color helpers
Move the frozen byte HSV and hex helpers behind a dependency-free module so the always-visible logo does not load Culori's parser and converter graph. Add a source-map build gate that fails if Culori is ever bundled into the byte-helper chunk.
