+++ apps/web/src/lib/color-bytes.ts
@@ -1,0 +1,22 @@
+export type RgbTuple = [number, number, number];
+
+// The tuple API rounds after adding HSV's value offset. Culori performs the
+// equivalent normalized-channel operations in another order, which changes
+// frozen byte results at half ties.
+export const hsvToRgb = (h: number, s: number, v: number): RgbTuple => {
+  const c = v * s;
+  const hp = h / 60;
+  const x = c * (1 - Math.abs((hp % 2) - 1));
+  let r = 0, g = 0, b = 0;
+  if (hp >= 0 && hp < 1) [r, g, b] = [c, x, 0];
+  else if (hp < 2) [r, g, b] = [x, c, 0];
+  else if (hp < 3) [r, g, b] = [0, c, x];
+  else if (hp < 4) [r, g, b] = [0, x, c];
+  else if (hp < 5) [r, g, b] = [x, 0, c];
+  else [r, g, b] = [c, 0, x];
+  const m = v - c;
+  return [Math.round((r + m) * 255), Math.round((g + m) * 255), Math.round((b + m) * 255)];
+};
+
+export const rgbToHex = (r: number, g: number, b: number): string =>
+  `#${[r, g, b].map(n => n.toString(16).padStart(2, '0')).join('').toUpperCase()}`;
